Reorganized components and directives into components module, run and config functions into files separate from app.js.
